# swagger.yaml
openapi: 3.0.0
info:
  title: CIS API
  description: API para gestionar temas, ideas y votación, integrando un sistema heredado.
  version: 1.0.0
servers:
  - url: http://localhost:8080/api
    description: Servidor local de desarrollo
paths:
...
  /topics/{id}:
    get:
      summary: Obtener detalles de un tema
      description: Obtiene los detalles de un tema específico por su ID.
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Detalles del tema.
        '404':
          description: Tema no encontrado.

    put:
      summary: Actualizar un tema
      description: Actualiza un tema existente.
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
      responses:
        '200':
          description: Tema actualizado exitosamente.
        '401':
          description: No autorizado (usuario no autenticado).
        '404':
          description: Tema no encontrado.

    delete:
      summary: Eliminar un tema
      description: Elimina un tema existente.
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Tema eliminado exitosamente.
        '401':
          description: No autorizado (usuario no autenticado).
        '404':
          description: Tema no encontrado.
...